+++ mindbug-backend/src/main/java/com/mindbug/models/Card.java
@@ -1,4 +1,4 @@
-package com.mindbug.model;
+package com.mindbug.models;
 
 import java.util.Set;
